Initial Work on Category Routes
12/14 Tests Passing

There is an issue with the test for the DELETE route. It is
passing tests, but the record is not deleted from the database.

The researched methods of handling record updates with the
Sequelize ORM are not working thus far.

Next Steps: Resolve issues with the DELETE route test and correct
issues with the DELETE and PUT controllers for categories.