Refactor API Routes to Nest Resources under Users
Several of the API routes are tied directly to a specific user.
As such, this commit refactors routes into a nested structure:

/api/blogs becomes /api/users/:userid/blogs
/api/jobs becomes /api/users/:userid/jobs
etc.

The folder structure of resources has been modified as well to
reflect this nested structure. The only resource the remains at
the root of the api alongside users is the categories route.