Remove Database Seed File for Testing
While considering ways to organize and run the test suite, I attempted
to create a seeding file that would add ProjectTypes to the database.
This file would run initially, but the database table was empty by the
time the tests for Projects would run. There must be some background
table sync process that happens during testing that I'm not aware of.

Regardless, as such this file is no longer needed.