Configure Susy Framework and Create About Partial
This project utilizes the Susy Framework for grid-based layouts.

To configure Susy properly, we update the package.json to include
Susy as a development dependency, and include an @import statement
for the framework in our main.scss stylesheet.

In _base.scss we begin to define the grid parameters for our
markup. On the home page <section> tags will be used to separate
page areas. Inside those sections, <article> tags will occupy
16 columns of our 24 column grid, with automatic left and right
margins.

In this commit, we add additional typographic styles, including
design elements for secondary headings. We also add the content
for the About section of the home page in a partial with an
accompanying Controller.