Reorganize Client Assets
Each major module in this application will be reconfigured such
that it operates as an Angular application (ng-app).

The blog.html file is being moved to /blog/index.html.

In addition, application-level files are being moved to
assets/js/modules.

There is an issue with navigation with this reconfiguration.
Pages do not load on click, but will load after a click and a
hard refresh.