Add Recent Posts and Global Footer to Home Page
Create a Blog Controller that contains a basic header.

Create a Footer Controller that calls the Profile service
to retrieve social media profile data used to build the
social icons at the bottom of the footer.

Add Sass to style the social icons and adjust global
typography settings as needed.

Create the Profile service used in the Footer controller.
Include an API call to /users/:id/profiles to retrieve all
of the social media profiles stored for the user's account.
Display them in the FooterController if the visible property
is set to true.