Convert Footer on Home Page to Global Control
The footer is one component that can be shared amongst all pages
in this application. Towards that end, we move the Footer
Controller from the _/js/controllers/home directory to a new
_/js/controllers/global directory.

In addition, we update the reference to this file in the  main
index.html file and update the Home application module.