Add the Global Footer to the Blog Page
Updates the Blog module’s index.html page to include a <footer>
element with a .global class to the bottom of the <div> that
contains the ng-app.

Adds references to the global FooterController and the Profile
service to the Blog index.html page.

Updates the Blog Angular module to include a reference to the
footerView in the blog $stateProvider.