Add the Global Footer to the Contact Page
Updates the contact module’s index.html page to include a <footer>
element with a .global class to the bottom of the <div> that
contains the ng-app.

Adds references to the global FooterController and the Profile
service to the Contact index.html page.

Updates the Contact Angular module to include a reference to the
footerView in the contact $stateProvider. Includes references to
the global.footer and profile.service Angular modules.