Add the Global Footer to the Portfolio Page
Updates the portfolio module’s index.html page to include a
<footer> element with a .global class to the bottom of the <div>
that contains the ng-app directive.

Adds references to the global FooterController and the Profile
service to the Portfolio index.html page.

Updates the Portfolio Angular module to include a reference to
the footerView in the portfolio $stateProvider. Includes references
to the global.footer and profile.service Angular modules.