Add the Global Footer to the Résumé Page
Updates the résumé module’s index.html page ot include a <footer>
element with a .global class to the bottom of the <div> that
contains the ng-app directive.

Adds references to the global FooterController and the Profile
service to the Résumé index.html page.

Updates the Resume Angular module to include a reference to the
footerView in the resume $stateProvider. Includes references to
the global.footer and profile.service Angular modules.