Set Up Resume Module
Create a separate stylesheet to manage styles for the Résumé page.
In _/sass/ create a resume.scss file that contains @import
statements to include all partials located in _/sass/resume.

Create partials for the primary sections of the Resume page: an
identifying header, a main content panel, and a sidebar.

Update the Resume module to include references to new controllers,
services and partials.

Test the page layout by adding section headings in the main
section and aside.