Finalize the Portfolio Module
Add a directory in the Sass repository to hold Porfolio module
styles (_/sass/portfolio). Create a partial for the gallery
styles and add style definitions for portfolio articles, headers
and images.

Update the client/portfolio/index.html page to include a <link>
tag that imports the Portfolio stylesheet. Also adjust the ui-view
name of the gallery <section> on line 26 (was: portfolioAboutView,
changed to: portfolioGalleryView).

Update the database seeding process to make portfolio items
generic, and adjust seeded image size for gallery items.