Update API Test Suite
Complete Tests and API Controller for Blog Categories.

Several of the tests in 13_blogcategories.js are failing.
It seems as if they are not even running. For example, the
first test in this document starting on line 11 returns an
empty object, although it is written exactly the same as
tests that complete the same action immediately following
the test. After the tests that post Blog Categories,
all remaining tests in the file fail.

Switching back to development mode to test on the
command line using HTTPie. Will enter a GitHub issue to
resolve this at a later date.