Add Client-side Pagination to Blogs Service
Update the getBlogPosts function in the Blog service so that it
includes two parameters: limit and offset with default values
in place (limit=3, offset=0).

Update the url to utilize these parameters.

In the BlogController, return posts.rows instead of posts.
The Sequelize findAndCountAll method returns an object with
two properties: count (an integer representing the total
number of posts) and rows (the actual blog post data).