Implement Category Filter for Blog Posts on Blog Index
In this commit, a solution for category filtering on Blog posts
is finalized. The major issue with the previous implementation was
that the filterOnName variable previously attached to $scope
was not being passed around to the controllers generated for the
various ui-views on the page (using Angular UI-Router). When a
category was clicked, it only updated the local $scope variable in
that container's scope.

The solution was to add the filterOnName variable to $rootScope,
revise the index.html page and associated partials so that there
was only one registration for the ng-controller directive
BlogController.

In addition, this commit implement other AngularJS features
including:

* the use of an ng-pluralize directive to use the
  appropriate plural form on the word 'post' based on the number of
  posts in each category

* the use of an alias in the ng-repeat directive to allow access to
  the returned filtered collection and its properties

* The addition of an anchor tag to the blog/index.html page
  called #top that can be used to scroll to the top of the page
  when a user clicks a category in the sidebar that is "below the
  fold". This required manipulation of the URL string so that the
  #top anchor reference does not persist in the address bar,
  preventing subsequent category clicks from scrolling to top