Update Anchor Tags for Blog Post Categories
At the bottom of each blog post there is a <footer> that contains
some metadata, including the date the post was published and
all associated categories.

In the initial implementation of this footer, we called a
function that did not properly implement the filter.

Here, we update the anchor tags for each category so that they
perform the same function as clicking a category link in the
sidebar.

On click, the page will filter by the clicked category, and the
focus of the page will be moved to the #top anchor tag so that
the user can see the selected category heading and the number of
posts in the category.