Remove BrowserSync Scripts from Index pages
BrowserSync was included as a part of the development process in
order to have the browser automatically reload on file changes.

The script throws a lot of warnings, and seems unnecessary at this
stage now that the majority of the front end components are complete.

To re-enable BrowserScript on any page, add the following:

  <script id="__bs_script__">
    //<![CDATA[
    document.write('<script async src=\'http://HOST:3000/browser-sync/browser-sync-client.js?v=2.18.12\'><\/script>'.replace('HOST', location.hostname));
    //]]>
  </script>