Add Dependencies: Milligram CSS Framework
The project needs a lightweight CSS Framework for form styling.
Bootstrap is popular, but more than what I'm looking for. As an
alternative, we will include a scaled-down version of the Milligram CSS
Framework (http://milligram.io/).

To get the core styles, we install milligram as a bower component. It
includes normalize - a popular CSS reset alternative.