Add Recent Blog Posts to the Home Page
Adds a function to the Blog Service called getLatestBlogPosts which
accepts a single parameter (limit) representing the number of blog posts
to retrieve (most recent first) in descending order.

Updates the Blog module to include a reference to the Blog service.

Updates the recent-blog-post partial to consume posts using the
getLatestBlogPosts method on the Blog service. Displays them using
minimalist typography and whitespace with large headers/titles for
readability.

Also included in this update is another refactor in Sass variables.
Now all typefaces have been included in the variables file (including
a new serif font selected for Blog posts) as well as
backgrounds/textures.

Closes #27, Closes #4