Database Schema Changes for Blog Posts
Adds a field 'teaser' to the Blog model. This field is used on the Home
Page when displaying recent posts.

Updates the seeding file to include teasers for the three most recent
blog posts.

Updates seedHelpers to include an image declaration in Markdown to
confirm that the markdown parser works for images.