Create Tests for Contact Page
The contact page should post a message to the server. What will happen
with that message will be determined at a later point... I think I want
to send the messages to my e-mail inbox or I plan to try to text them to
my phone. In the interim, here we write a simple test that confirms that
the messages POSTed to /api/messages are received by the server as a
JSON object.