Handle Processing of Contact Form
Closes #8

Add bare minimum updates to client-side code in order to handle contact
form processing.

The Message service includes a single function (sendMessage) that
executes a POST request on the route /api/messages, sending passed in
message object to the server. The contct module was updated to include
this service.

The FormController was updated to include a defaultForm variable which
can be used to clear the form values after a form is submitted. In
addition, we add a sendMessage function to the $scope, which takes the
fom data and sends it to the server using the Message service's
sendMessage function. After sending the message, the function sets the
form back to a pristine state, sets the form fields to the defaultForm
values, and sets the form to an untouched state.

The form partial was updated and simplified. A form name and ng-submit
directive were added, ng-odel attributes were added to each form field.

The client/index.html page was updated to include a <script> tag that
includes the message service.