Corrects the GET Limited Posts Route
(Resolves #43)

The offset value passed to this route was configured with an incorrect
parameter. To correct this, we assign the offset variable to equal the
offset parameter (it was calculating this value based on the page, which
no longer exists).

In addition, we return offset and limit as Integer values instead of
strings with the number of pages in the collection based on the limit.

Removes the page (current page number) metadata property from the response object.