Updates /api/blogs/posts/:limit/:offset
(Resolves #43)
Updates the blog posts API route that handles pagination.

Adds the limit and offset parameters to the returned JSON object.
Adds a calculated page count based on the offset and limit.
Returns the current page in the JSON object.
Returns an array containing the page numbers for the collection in
the JSON object.