Update Main Blog Controller
Convert all $scope variables related to pagination and filtering to
$rootScope variables.

Modifies the routine that retrieves Blog Posts from the API such that it
retrieves the count of all posts first, so that variable can be passed
to the call to .getLatestBlogPosts.

Modifies the routine that filters blog posts by category such that it
assigns posts to posts.rows (the API has been modified to use the
Sequelize findAndCountAll function) and updates the count and pages
$rootScope variables.

Adds a globally-scoped variable named postCount.