Update Blog Partials
In main.html - the partial that controls the display of blog posts on
the main Blog page - add back the limitTo setting.

In pagination.html - remove the test "print to screen" data for the
number of blog posts. Implement the paging control to display a button
for each page in the collection. Display the number of post displaying
and the number of total posts (Posts x - y of z).