Update the /api/blogs/categories Controller
Modify the one function - it returns all Blog posts that belong to a
given category - such that it calls the findAndCountAll sequelize
method instead of findAll.

Update the data returned from the controller such that it includes an
array of the number of pages (given each page contains three items) in
the collection.