Create Responsive Top Navigation
Resolves #51

Removes unnecessary partial:
  client/assets/partials/global/navigation.html

Updates navigation markup in each index.html file to include a .main
class on the <nav> tag, and add a checkbox and label element to
control the visibility of the hamburger menu icon and manage the
transition of the menu when the page is resized.

Adds media queries and markup to turn the navigation menu from a
static horizontal menu to a responsive menu.

CREDIT: Inspired by a Pure CSS demonstration of "hamburger menu"
navigation on CodePen (https://codepen.io/torbencolding/pen/OPwwOg).