Add Logo to Top Navigation Bar
Resolves #55

After initial mockups with the original circular logo design, I changed
the logo to a square, outlined logo with text. This logo has been added
to the top navigation bar in each index.html page, and the top
navigation styles have been adjusted to accomodate it. The logo position
adjusts based on the viewport width.