Adjust H2::after Effect to Correct Overlay Issue
When the top navigation transitions from wide with a horizontal menu to
narrow/responsive with a "hamburger menu" the ::after effect tied to H2
elements in the <header> has a high enough z-index to appear on top of
the navigation menu.

Adjust the width at which this element is positioned so that it moves
behind the menu when in responsive/mobile layout mode but adjusts as
desired on wide screens.